+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.0 / 2024-12-31
 
-* Now replaces include variables with their values
+* Now replaces include and layout variables with their values
 
 
 ## 1.0.3 / 2024-08-19